Product Quantity needs to be changed back to a decimal from an Int

+++ modules/AOS_Products_Quotes/vardefs.php
@@ -112,7 +112,7 @@ $dictionary['AOS_Products_Quotes'] = array(
     'required' => false,
     'name' => 'product_qty',
     'vname' => 'LBL_PRODUCT_QTY',
-    'type' => 'integer',  // it was: 'type' => 'decimal',
+    'type' => 'decimal',
     'massupdate' => 0,
     'comments' => '',
     'help' => '',
